requirements.txt
langchain
chromadb
openai
pypdf
python-dotenv
langchain-google-genai
google-generativeai
sentence-transformers
streamlit
